Install the given version of sops
To install the sops binary we are using a package from github called
tools-cache which donwloads the packages and stores them at a particular
location in the runner and re-uses these packages whenever required
instead of downloading them again from internet for other workflow runs.

As the sops binary has to be downloaded from the internet, it takes some
amount of time to execute the tests also.

REF:
https://github.com/actions/toolkit/tree/master/packages/tool-cache