Get correct download link based on platform
Based on the platform/runner selected in the github workflow we should
be able to get the correct download link for sops.
Also, refactored the code a little bit in order to avoid stuttering in
the function names.

Along with these, we now have two npm commands to execute different
types of tests.
`npm run unitTests` executes the unit tests. Similarly,
`npm run journeyTests` executes the journey tests.