run all tests and format the code with npm scripts
From a developers point of view, the question after having performed
changes, is "do the tests pass". And not so much "do the unit tests
pass" and then "do the journey tests pass", while having the ability to
run those separately in some instances might be useful, it shouldn't be
the norm.