Throw an error when importing gpg key fails
whenever the execution of gpg command fails we are returning the error
output as is to the calling function where we can later handle the
execution of flow of action.

The testing of this type of functions which throws an error was a bit
tricky. I used the following refrence to solve the testing part of it.
http://eloquentcode.com/expect-a-function-to-throw-an-exception-in-jest