Set output variable of action with decrypted data
As of now, SOPS decrypts the data and outputs it as JSON format only.
Later, we can make this configurable if needed.
Also, for decryption we are using the absolute path of sops binary where
it is installed in order to avoid using a different version of SOPS if
any exists already on the runner.