Delete the gpg key imported after decryption
We have to clean the GPG key imported that is used to decrypt the
encrypted file. Importing a private GPG key results in the import of
both private and public keys to the keyring. To delete these we have to
first delete the private key and then the public key.