Delete gpg key if it exists in post action step
If the sops-decrypt action is used multiple times in a single workflow
job using the same private gpg key to decrypt then we will have only one
key imported for this workflow job in the runner.

As the action is used mutliple times the post step is also triggered
multiple times at the end of the workflow job. The execution of first
post step gets succeeded by deleting the key and the rest fails if we
don't check for the key existence as the key is already deleted in the
first post step execution.