Remove sonarqube scans for now
The sonarqube project which is being used is internal which needs a
self-hosted runner to scan the repository. We cannot use self-hosted
runners on public repositories as it is a security risk as stated in the
GitHub documentation. We need to find a solution for this before we can
assign a self-hosted runner.
Ref: https://docs.github.com/en/actions/hosting-your-own-runners/about-self-hosted-runners#self-hosted-runner-security-with-public-repositories