Refactoring and Node16
Refactor Codebase and set Node16 as default runtime

Updating the required nodejs version for the github-action to node16.
Dependencies and devDependencies were updated to the latest versions. 
Various dependencies were removed, e.g. prettier, semantic release, as they are not necessary.
For better performance a node 16 specific tsconfig.json is now used.
ncc was replaced by esbuild, as it is faster and the more modern approach.
The build pipeline was simplified and extended, for improved confidence regarding updates coming from dependabot.
A build-action.yaml was added to allow transpiling the bot in a github workflow. Thus making it easier to patch dependabot pull-requests if they affect the generated code.